backend/requirements.txt
Django >= 3.2, < 3.3.0
psycopg2-binary >= 2.9, < 3.0.0
ptvsd >= 4.3, < 4.4.0
python-decouple >= 3.5, < 3.6.0
gunicorn >= 20.1.0, < 20.2.0
whitenoise >= 5.3.0, < 5.4.0
Brotli >= 1.0.9, < 1.1.0
dj-database-url >= 0.5.0, < 0.6.0
djangorestframework >= 3.13.1, < 3.14.0
